requirements.txt
python-dotenv>=1.0.0
SQLAlchemy>=2.0
PyMySQL>=1.1
